refactor: split electronic warfare systems by concern
