+++ src/utils/gameplay/electronicWarfare/constants.ts
@@ -1,0 +1,13 @@
+export const ECM_RADIUS = 6;
+
+export const BAP_ECM_COUNTER_RANGE = 4;
+
+export const BLOODHOUND_ECM_COUNTER_RANGE = 8;
+
+export const CLAN_PROBE_ECM_COUNTER_RANGE = 5;
+
+export const STEALTH_ARMOR_MODIFIERS = {
+  short: 0,
+  medium: 1,
+  long: 2,
+} as const;
